+++ backend/models/match_model.py
@@ -2,10 +2,11 @@ from pydantic import BaseModel
 from typing import Optional
 from datetime import datetime
 from uuid import UUID
+from decimal import Decimal
 
 class MatchRequest(BaseModel):
     corridor: str
-    amount: float
+    amount: Decimal
     payment_method: str
     delivery_speed: str
     beneficiary_country: str
@@ -14,7 +15,7 @@ class MatchRequest(BaseModel):
 class MatchResponse(BaseModel):
     uuid: UUID
     counterparty_uuid: UUID
-    market_rate_used: float
+    market_rate_used: Decimal
     rate_source: str
     rate_fallback: bool
     rate_reason: str
